hmi and mmpoc demo: Remove meta-qt5 from template
Before we add meta-qt5 to these demo to prevent warning message
(as there are some orphan .bbappend for qt).
But this can cause problem now as meta-qt5 has updated, and not
compatible with current poky version. On Qt and Doorphone manual
we instruction user to checkout an old version for meta-qt5.

It is not suitable for HMI and MMPOC as these demos do not use Qt
at all. So remove it here and in manual, inform user to ignore warning.

Signed-off-by: Hung Tran <hung.tran.jy@rvc.renesas.com>