qt-hmi-demo: Change to GPU mode for SKRZG1M
On current environment for SKRZG1M, 3D and Qt performance
in VSP mode is lower than GPU mode.
Thus this patch temporarily change default mode for qt-hmi-demo
to GPU mode instead.
Note that this change cannot apply to SKRZG1E because
on SKRZG1E GPU mode performance is still lower than VSP

The workaround "revert update due to low performance is
also removed as it is not neccessary anymore.

Signed-off-by: HungTran <hung.tran.jy@rvc.renesas.com>